requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/base.txt --strip-extras requirements/base.in
#
annotated-types==0.7.0
    # via pydantic
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
asttokens==3.0.0
    # via stack-data
bertviz==1.4.0
    # via -r requirements/base.in
boto3==1.37.13
    # via bertviz
botocore==1.37.13
    # via
    #   boto3
    #   s3transfer
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
comm==0.2.2
    # via ipywidgets
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
decorator==5.2.1
    # via ipython
executing==2.2.0
    # via stack-data
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fonttools==4.56.0
    # via matplotlib
fsspec==2025.3.0
    # via
    #   huggingface-hub
    #   torch
huggingface-hub==0.29.3
    # via
    #   tokenizers
    #   transformers
hydra-core==1.3.2
    # via -r requirements/base.in
idna==3.10
    # via requests
ipython==8.34.0
    # via
    #   -r requirements/base.in
    #   ipywidgets
ipywidgets==8.1.5
    # via rich
iso639-lang==2.6.0
    # via -r requirements/base.in
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via torch
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.4.2
    # via scikit-learn
jupyterlab-widgets==3.0.13
    # via ipywidgets
kiwisolver==1.4.8
    # via matplotlib
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.1
    # via -r requirements/base.in
matplotlib-inline==0.1.7
    # via ipython
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==1.26.4
    # via
    #   -r requirements/base.in
    #   contourpy
    #   matplotlib
    #   scikit-learn
    #   scipy
    #   transformers
omegaconf==2.3.0
    # via hydra-core
packaging==24.2
    # via
    #   huggingface-hub
    #   hydra-core
    #   matplotlib
    #   transformers
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==11.1.0
    # via matplotlib
prompt-toolkit==3.0.50
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pydantic==2.10.6
    # via -r requirements/base.in
pydantic-core==2.27.2
    # via pydantic
pygments==2.19.1
    # via
    #   ipython
    #   rich
pyparsing==3.2.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   matplotlib
pyyaml==6.0.2
    # via
    #   -r requirements/base.in
    #   huggingface-hub
    #   omegaconf
    #   transformers
regex==2024.11.6
    # via
    #   -r requirements/base.in
    #   bertviz
    #   transformers
requests==2.32.3
    # via
    #   -r requirements/base.in
    #   bertviz
    #   huggingface-hub
    #   transformers
rich==13.9.4
    # via -r requirements/base.in
s3transfer==0.11.4
    # via boto3
safetensors==0.5.3
    # via transformers
scikit-learn==1.6.1
    # via -r requirements/base.in
scipy==1.15.2
    # via scikit-learn
sentencepiece==0.2.0
    # via bertviz
six==1.17.0
    # via python-dateutil
stack-data==0.6.3
    # via ipython
sympy==1.13.3
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.1
    # via transformers
torch==2.2.2
    # via
    #   -r requirements/base.in
    #   bertviz
torchinfo==1.8.0
    # via -r requirements/base.in
tqdm==4.67.1
    # via
    #   -r requirements/base.in
    #   bertviz
    #   huggingface-hub
    #   transformers
traitlets==5.14.3
    # via
    #   comm
    #   ipython
    #   ipywidgets
    #   matplotlib-inline
transformers==4.49.0
    # via bertviz
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   ipython
    #   pydantic
    #   pydantic-core
    #   torch
urllib3==2.3.0
    # via
    #   botocore
    #   requests
wcwidth==0.2.13
    # via prompt-toolkit
widgetsnbextension==4.0.13
    # via ipywidgets
